Add files via upload
This dataset offers a glimpse into data science job salaries from 2020 to 2024, detailing experience levels, employment types, job titles, and company attributes. It's a valuable resource for understanding salary trends and factors shaping compensation in the data science realm.
This dataset provides insights into data science job salaries from 2020 to 2024, including information on experience levels, employment types, job titles, and company characteristics. It serves as a valuable resource for understanding salary trends and factors influencing compensation in the data science field.

Features:
work_year: The year of the data related to the job salary.
experience_level: The level of experience of the employee (e.g., entry-level, mid-level, senior-level).
employment_type: The type of employment (e.g., full-time, part-time, contract).
job_title: The title or role of the employee within the data science field.
salary: The salary of the employee.
salary_currency: The currency in which the salary is denoted.
salary_in_usd: The salary converted to US dollars for standardization.
employee_residence: The residence location of the employee.
remote_ratio: The ratio of remote work allowed for the position.
company_location: The location of the company.
company_size: The size of the company based on employee count or revenue.